fix(rebase): rebase on deser changes
