+++ configuration/src/chains/ethereum/submitter/gelato.rs
@@ -5,16 +5,16 @@ use crate::{agent::SignerConf, FromEnv};
 #[serde(rename_all = "camelCase")]
 pub struct GelatoConf {
     /// Sponsor signer configuration
-    pub signer: SignerConf,
+    pub sponsor: SignerConf,
     /// Address of fee token
     pub fee_token: String,
 }
 
 impl FromEnv for GelatoConf {
     fn from_env(prefix: &str) -> Option<Self> {
-        let signer = SignerConf::from_env(&format!("{}_SIGNER", prefix))?;
+        let sponsor = SignerConf::from_env(&format!("{}_SPONSOR", prefix))?;
         let fee_token = std::env::var(&format!("{}_FEETOKEN", prefix)).ok()?;
 
-        Some(Self { signer, fee_token })
+        Some(Self { sponsor, fee_token })
     }
 }
