+++ configuration/src/chains/mod.rs
@@ -23,9 +23,11 @@ impl Default for ChainConf {
 }
 
 impl FromEnv for ChainConf {
-    fn from_env(prefix: &str) -> Option<Self> {
-        let rpc_style = std::env::var(&format!("{}_RPCSTYLE", prefix)).ok()?;
-        let rpc_url = std::env::var(&format!("{}_CONNECTION_URL", prefix)).ok()?;
+    fn from_env(network: &str) -> Option<Self> {
+        println!("Getting chainconf info");
+        let rpc_style = std::env::var(&format!("RPCS_{}_RPCSTYLE", network)).ok()?;
+        let rpc_url = std::env::var(&format!("RPCS_{}_CONNECTION_URL", network)).ok()?;
+        println!("GOT chainconf info");
 
         let json = json!({
             "rpcStyle": rpc_style,
