+++ fixtures/env.test
@@ -26,8 +26,7 @@ TRANSACTIONSUBMITTERS_MOONBEAM_SUBMITTER_FEETOKEN=0xabcd
 
 TXSUBMITTERS_MOONBEAM_RPCSTYLE=ethereum
 TXSUBMITTERS_MOONBEAM_SUBMITTERTYPE=gelato
-TXSUBMITTERS_MOONBEAM_SUBMITTER_SIGNER_KEYTYPE=hexKey
-TXSUBMITTERS_MOONBEAM_SUBMITTER_SIGNER_KEY=0x1111111111111111111111111111111111111111111111111111111111111111
+TXSUBMITTERS_MOONBEAM_SUBMITTER_SPONSOR_KEY=0x1111111111111111111111111111111111111111111111111111111111111111
 TXSUBMITTERS_MOONBEAM_SUBMITTER_FEETOKEN=0xabcd
 
 TRANSACTIONSIGNERS_EVMOS_KEY=0x1111111111111111111111111111111111111111111111111111111111111112
