+++ fixtures/test_secrets.json
@@ -14,19 +14,13 @@
     "ethereum": {
       "rpcStyle": "ethereum",
       "submitterType": "local",
-      "submitter": {
-        "key": "0x1111111111111111111111111111111111111111111111111111111111111111",
-        "keyType": "hexKey"
-      }
+      "submitter": "0x1111111111111111111111111111111111111111111111111111111111111111"
     },
     "moonbeam": {
       "rpcStyle": "ethereum",
       "submitterType": "gelato",
       "submitter": {
-        "signer": {
-          "key": "0x1111111111111111111111111111111111111111111111111111111111111111",
-          "keyType": "hexKey"
-        },
+        "sponsor": "0x1111111111111111111111111111111111111111111111111111111111111111",
         "feeToken": "0xabc"
       }
     }
